--- Sources/ACKLocalizationCore/Model/PluralKey.swift
@@ -1,13 +1,0 @@
-//
-//  PluralKey.swift
-//  
-//
-//  Created by Lukáš Hromadník on 07/07/2020.
-//
-
-import Foundation
-
-struct PluralKey: Codable {
-    let key: PluralTranslationKey
-    let value: String
-}
